separate test for miscNonRentExempt that uses old validator
