+++ .github/workflows/tests.yaml
@@ -239,6 +239,41 @@ jobs:
       - run: cd tests/bpf-upgradeable-state && cp bpf_upgradeable_state-keypair.json target/deploy/bpf_upgradeable_state-keypair.json && anchor test --skip-local-validator --skip-build --skip-lint
       - uses: ./.github/actions/git-diff/
 
+  # this test exists to make sure that anchor
+  # checks rent correctly for legacy accounts
+  # that don't have to be rent-exempt
+  test-misc-non-rent-exempt:
+    needs: setup-anchor-cli # this is built with a different solana version but that's fine since it's just the cli
+    name: Test tests/misc/nonRentExempt
+    runs-on: ubuntu-18.04
+    timeout-minutes: 30
+    steps:
+      - uses: actions/checkout@v2
+      - uses: ./.github/actions/setup/
+      - uses: ./.github/actions/setup-ts/
+      - uses: actions/cache@v2
+        name: Cache Solana Tool Suite
+        id: cache-solana
+        with:
+          path: |
+            ~/.cache/solana/
+            ~/.local/share/solana/
+          key: solana-${{ runner.os }}-v0000-1.8.14
+      # using an outdated validator but that is ok as long as the test doesn't include newer incompatible features
+      - run: sh -c "$(curl -sSfL https://release.solana.com/v1.8.14/install)"
+        shell: bash
+      - run: echo "/home/runner/.local/share/solana/install/active_release/bin" >> $GITHUB_PATH
+        shell: bash
+      - run: solana-keygen new --no-bip39-passphrase
+        shell: bash
+      - run: solana config set --url localhost
+        shell: bash
+      - run: cd tests/misc && yarn --frozen-lockfile
+      - run: cd tests/misc && yarn link @project-serum/anchor
+      - run: cd tests/misc && anchor build --skip-lint
+      - run: solana-test-validator --bpf-program Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS tests/misc/target/deploy/misc.so -r --quiet &
+      - run: cd tests/misc && anchor run test_non_rent_exempt
+
   test-anchor-init:
     needs: setup-anchor-cli
     name: Test Anchor Init
